PPU improvements. Scrolling works, and so does SMB!

+++ src/main/java/com/zacharyhirsch/moldynes/emulator/ppu/NesPpuSprite.java
@@ -1,6 +1,6 @@
 package com.zacharyhirsch.moldynes.emulator.ppu;
 
-record NesPpuSprite(int index, int y, int tileIndex, int attributes, int x) {
+record NesPpuSprite(int index, int y, int tileIndex, byte attributes, int x) {
 
   public boolean valid() {
     return y() != 0xff && x() != 0xff;
